+++ src/main/java/lotto/domain/exception/DuplicateLottoNumberException.java
@@ -1,4 +1,4 @@
-package lotto.exception;
+package lotto.domain.exception;
 
 public class DuplicateLottoNumberException extends IllegalArgumentException {
     public DuplicateLottoNumberException(String message) {
