+++ src/main/java/lotto/domain/exception/InvalidCountOfLottoNumberException.java
@@ -1,4 +1,4 @@
-package lotto.exception;
+package lotto.domain.exception;
 
 public class InvalidCountOfLottoNumberException extends IllegalArgumentException {
     public InvalidCountOfLottoNumberException(String message) {
