+++ src/main/java/lotto/domain/exception/NaturalNumberException.java
@@ -1,4 +1,4 @@
-package lotto.exception;
+package lotto.domain.exception;
 
 public class NaturalNumberException extends IllegalArgumentException {
     public NaturalNumberException(String message) {
